Move manifest factory methods to ManifestFiles.
