+++ core/src/main/java/org/apache/iceberg/ManifestReader.java
@@ -45,8 +45,7 @@ import static org.apache.iceberg.expressions.Expressions.alwaysTrue;
 /**
  * Reader for manifest files.
  * <p>
- * The preferable way to create readers is using {@link #read(ManifestFile, FileIO, Map)} as
- * it allows entries to inherit manifest metadata such as snapshot id.
+ * Create readers using {@link ManifestFiles#read(ManifestFile, FileIO, Map)}.
  */
 public class ManifestReader extends CloseableGroup implements Filterable<FilteredManifest> {
   private static final Logger LOG = LoggerFactory.getLogger(ManifestReader.class);
@@ -62,61 +61,16 @@ public class ManifestReader extends CloseableGroup implements Filterable<Filtere
   /**
    * Returns a new {@link ManifestReader} for an {@link InputFile}.
    * <p>
-   * <em>Note:</em> Most callers should use {@link #read(ManifestFile, FileIO, Map)} to ensure that
+   * <em>Note:</em> Callers should use {@link ManifestFiles#read(ManifestFile, FileIO, Map)} to ensure that
    * manifest entries with partial metadata can inherit missing properties from the manifest metadata.
-   * <p>
-   * <em>Note:</em> Most callers should use {@link #read(InputFile, Map)} if all manifest entries
-   * contain full metadata and they want to ensure that the schema used by filters is the latest
-   * table schema. This should be used only when reading a manifest without filters.
    *
    * @param file an InputFile
    * @return a manifest reader
+   * @deprecated use {@link ManifestFiles#read(ManifestFile, FileIO, Map)}.
    */
+  @Deprecated
   public static ManifestReader read(InputFile file) {
     return new ManifestReader(file, null, InheritableMetadataFactory.empty());
-  }
-
-  /**
-   * Returns a new {@link ManifestReader} for an {@link InputFile}.
-   * <p>
-   * <em>Note:</em> Most callers should use {@link #read(ManifestFile, FileIO, Map)} to ensure that
-   * manifest entries with partial metadata can inherit missing properties from the manifest metadata.
-   *
-   * @param file an InputFile
-   * @param specsById a Map from spec ID to partition spec
-   * @return a manifest reader
-   */
-  public static ManifestReader read(InputFile file, Map<Integer, PartitionSpec> specsById) {
-    return new ManifestReader(file, specsById, InheritableMetadataFactory.empty());
-  }
-
-  /**
-   * Returns a new {@link ManifestReader} for a {@link ManifestFile}.
-   * <p>
-   * <em>Note:</em> Most callers should use {@link #read(ManifestFile, FileIO, Map)} to ensure
-   * the schema used by filters is the latest table schema. This should be used only when reading
-   * a manifest without filters.
-   *
-   * @param manifest a ManifestFile
-   * @param io a FileIO
-   * @return a manifest reader
-   */
-  public static ManifestReader read(ManifestFile manifest, FileIO io) {
-    return read(manifest, io, null);
-  }
-
-  /**
-   * Returns a new {@link ManifestReader} for a {@link ManifestFile}.
-   *
-   * @param manifest a ManifestFile
-   * @param io a FileIO
-   * @param specsById a Map from spec ID to partition spec
-   * @return a manifest reader
-   */
-  public static ManifestReader read(ManifestFile manifest, FileIO io, Map<Integer, PartitionSpec> specsById) {
-    InputFile file = io.newInputFile(manifest.path());
-    InheritableMetadata inheritableMetadata = InheritableMetadataFactory.fromManifest(manifest);
-    return new ManifestReader(file, specsById, inheritableMetadata);
   }
 
   private final InputFile file;
@@ -129,7 +83,7 @@ public class ManifestReader extends CloseableGroup implements Filterable<Filtere
   private List<ManifestEntry> cachedAdds = null;
   private List<ManifestEntry> cachedDeletes = null;
 
-  private ManifestReader(InputFile file, Map<Integer, PartitionSpec> specsById,
+  ManifestReader(InputFile file, Map<Integer, PartitionSpec> specsById,
                          InheritableMetadata inheritableMetadata) {
     this.file = file;
     this.inheritableMetadata = inheritableMetadata;
