+++ core/src/test/java/org/apache/iceberg/TestManifestReader.java
@@ -30,6 +30,7 @@ import org.junit.Test;
 public class TestManifestReader extends TableTestBase {
 
   @Test
+  @SuppressWarnings("deprecation")
   public void testManifestReaderWithEmptyInheritableMetadata() throws IOException {
     ManifestFile manifest = writeManifest("manifest.avro", manifestEntry(Status.EXISTING, 1000L, FILE_A));
     try (ManifestReader reader = ManifestReader.read(FILE_IO.newInputFile(manifest.path()))) {
@@ -41,6 +42,7 @@ public class TestManifestReader extends TableTestBase {
   }
 
   @Test
+  @SuppressWarnings("deprecation")
   public void testInvalidUsage() throws IOException {
     ManifestFile manifest = writeManifest(FILE_A, FILE_B);
     try (ManifestReader reader = ManifestReader.read(FILE_IO.newInputFile(manifest.path()))) {
